* as.c (main): Move parse_args before symbol_begin and frag_init so that the   hash table size can be set before it is used. * hash.c: Use an unsigned long type for the size of the hash tables. * hash.h (set_gas_hash_table_size): Update the prototype.

+++ gas/hash.h
@@ -26,7 +26,7 @@ struct hash_control;
 
 /* Set the size of the hash table used.  */
 
-void set_gas_hash_table_size (unsigned int);
+void set_gas_hash_table_size (unsigned long);
 
 /* Create a hash table.  This return a control block.  */
